dashboard/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
kubernetes==31.0.0
jinja2==3.1.4
boto3==1.35.0
httpx==0.27.0
itsdangerous==2.2.0
python-jose[cryptography]==3.3.0
